8247474: Shenandoah: Windows build warning after JDK-8247310

+++ src/hotspot/share/gc/shenandoah/shenandoahPacer.hpp
@@ -101,7 +101,7 @@ private:
 
   size_t update_and_get_progress_history();
 
-  void wait(long time_ms);
+  void wait(size_t time_ms);
   void notify_waiters();
 };
